Added auto refresh functionality and disabled caching for API on browsers.
XMLHttpRequest was ditched in favor of Fetch.

Every 60s, district data is requested from the API. Every time the timer for an invasion ends, a request is also made after 5s.

District element ids were changed to their names so that elements could be replaced instead of reloading the webpage.